--- extension/manifest_cr.json
@@ -1,55 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "DarkWatt",
-  "short_name": "DarkWatt",
-  "description": "DarkWatt is a tool that helps you find the best dark mode for your website.",
-  "version": "0.1.0",
-  "author": "@Luca1905",
-  "offline_enabled": true,
-  "action": {
-      "default_title": "DarkWatt",
-      "default_icon": {
-          "256": "assets/icon256.png"
-      },
-      "default_popup": "js/popup.html"
-  },
-  "content_security_policy": {
-      "extension_pages": "script-src 'self' 'wasm-unsafe-eval'; object-src 'self';"
-  },
-  "background": {
-      "type": "module",
-      "service_worker": "js/background.js"
-  },
-  "content_scripts": [
-      {
-          "matches": [
-              "<all_urls>"
-          ],
-          "js": [
-              "js/content.js"
-          ],
-          "run_at": "document_end"
-      }
-  ],
-  "permissions": [
-      "webRequest",
-      "tabs",
-      "processes",
-      "system.display"
-  ],
-  "host_permissions": [
-      "<all_urls>"
-  ],
-  "web_accessible_resources": [
-      {
-          "matches": [
-              "<all_urls>"
-          ],
-          "resources": [
-              "js/utils/logger.js",
-              "js/wasm/wasm_mod.js",
-              "js/wasm/wasm_mod_bg.wasm"
-          ]
-      }
-  ]
-}
