+++ extension/public/icons/icon256.png
