+++ extension/src/utils/logger.js
@@ -1,3 +1,3 @@
 export const log = (tag, ...args) => console.log(`[${tag}]`, ...args);
 export const warn = (tag, ...args) => console.warn(`[${tag}]`, ...args);
-export const error = (tag, ...args) => console.error(`[${tag}]`, ...args); 
+export const error = (tag, ...args) => console.error(`[${tag}]`, ...args);
