+++ extension/src/wasm/wasm_mod.js
@@ -1,0 +1,388 @@
+let wasm;
+
+function logError(f, args) {
+	try {
+		return f.apply(this, args);
+	} catch (e) {
+		const error = (() => {
+			try {
+				return e instanceof Error
+					? `${e.message}\n\nStack:\n${e.stack}`
+					: e.toString();
+			} catch (_) {
+				return "<failed to stringify thrown value>";
+			}
+		})();
+		console.error(
+			"wasm-bindgen: imported JS function that was not marked as `catch` threw an error:",
+			error,
+		);
+		throw e;
+	}
+}
+
+const cachedTextDecoder =
+	typeof TextDecoder !== "undefined"
+		? new TextDecoder("utf-8", { ignoreBOM: true, fatal: true })
+		: {
+				decode: () => {
+					throw Error("TextDecoder not available");
+				},
+			};
+
+if (typeof TextDecoder !== "undefined") {
+	cachedTextDecoder.decode();
+}
+
+let cachedUint8ArrayMemory0 = null;
+
+function getUint8ArrayMemory0() {
+	if (
+		cachedUint8ArrayMemory0 === null ||
+		cachedUint8ArrayMemory0.byteLength === 0
+	) {
+		cachedUint8ArrayMemory0 = new Uint8Array(wasm.memory.buffer);
+	}
+	return cachedUint8ArrayMemory0;
+}
+
+function getStringFromWasm0(ptr, len) {
+	ptr = ptr >>> 0;
+	return cachedTextDecoder.decode(
+		getUint8ArrayMemory0().subarray(ptr, ptr + len),
+	);
+}
+
+export function hello_wasm() {
+	wasm.hello_wasm();
+}
+
+let WASM_VECTOR_LEN = 0;
+
+function passArray8ToWasm0(arg, malloc) {
+	const ptr = malloc(arg.length * 1, 1) >>> 0;
+	getUint8ArrayMemory0().set(arg, ptr / 1);
+	WASM_VECTOR_LEN = arg.length;
+	return ptr;
+}
+/**
+ * @param {Uint8Array} pixels
+ * @returns {number}
+ */
+export function average_luma_relative(pixels) {
+	const ptr0 = passArray8ToWasm0(pixels, wasm.__wbindgen_malloc);
+	const len0 = WASM_VECTOR_LEN;
+	const ret = wasm.average_luma_relative(ptr0, len0);
+	return ret;
+}
+
+/**
+ * @param {Uint8Array} pixels
+ * @returns {number}
+ */
+export function average_luma_in_nits(pixels) {
+	const ptr0 = passArray8ToWasm0(pixels, wasm.__wbindgen_malloc);
+	const len0 = WASM_VECTOR_LEN;
+	const ret = wasm.average_luma_in_nits(ptr0, len0);
+	return ret;
+}
+
+const cachedTextEncoder =
+	typeof TextEncoder !== "undefined"
+		? new TextEncoder("utf-8")
+		: {
+				encode: () => {
+					throw Error("TextEncoder not available");
+				},
+			};
+
+const encodeString =
+	typeof cachedTextEncoder.encodeInto === "function"
+		? (arg, view) => cachedTextEncoder.encodeInto(arg, view)
+		: (arg, view) => {
+				const buf = cachedTextEncoder.encode(arg);
+				view.set(buf);
+				return {
+					read: arg.length,
+					written: buf.length,
+				};
+			};
+
+function passStringToWasm0(arg, malloc, realloc) {
+	if (typeof arg !== "string")
+		throw new Error(`expected a string argument, found ${typeof arg}`);
+
+	if (realloc === undefined) {
+		const buf = cachedTextEncoder.encode(arg);
+		const ptr = malloc(buf.length, 1) >>> 0;
+		getUint8ArrayMemory0()
+			.subarray(ptr, ptr + buf.length)
+			.set(buf);
+		WASM_VECTOR_LEN = buf.length;
+		return ptr;
+	}
+
+	let len = arg.length;
+	let ptr = malloc(len, 1) >>> 0;
+
+	const mem = getUint8ArrayMemory0();
+
+	let offset = 0;
+
+	for (; offset < len; offset++) {
+		const code = arg.charCodeAt(offset);
+		if (code > 0x7f) break;
+		mem[ptr + offset] = code;
+	}
+
+	if (offset !== len) {
+		if (offset !== 0) {
+			arg = arg.slice(offset);
+		}
+		ptr = realloc(ptr, len, (len = offset + arg.length * 3), 1) >>> 0;
+		const view = getUint8ArrayMemory0().subarray(ptr + offset, ptr + len);
+		const ret = encodeString(arg, view);
+		if (ret.read !== arg.length) throw new Error("failed to pass whole string");
+		offset += ret.written;
+		ptr = realloc(ptr, len, offset, 1) >>> 0;
+	}
+
+	WASM_VECTOR_LEN = offset;
+	return ptr;
+}
+/**
+ * @param {string} uri
+ * @returns {number}
+ */
+export function average_luma_in_nits_from_data_uri(uri) {
+	const ptr0 = passStringToWasm0(
+		uri,
+		wasm.__wbindgen_malloc,
+		wasm.__wbindgen_realloc,
+	);
+	const len0 = WASM_VECTOR_LEN;
+	const ret = wasm.average_luma_in_nits_from_data_uri(ptr0, len0);
+	return ret;
+}
+
+function _assertNum(n) {
+	if (typeof n !== "number")
+		throw new Error(`expected a number argument, found ${typeof n}`);
+}
+/**
+ * @param {number} display_width
+ * @param {number} display_height
+ * @param {number} hours
+ * @param {DisplayTech} tech
+ * @param {string} uri
+ * @returns {number}
+ */
+export function estimate_saved_energy_mwh_from_data_uri(
+	display_width,
+	display_height,
+	hours,
+	tech,
+	uri,
+) {
+	_assertNum(display_width);
+	_assertNum(display_height);
+	_assertNum(tech);
+	const ptr0 = passStringToWasm0(
+		uri,
+		wasm.__wbindgen_malloc,
+		wasm.__wbindgen_realloc,
+	);
+	const len0 = WASM_VECTOR_LEN;
+	const ret = wasm.estimate_saved_energy_mwh_from_data_uri(
+		display_width,
+		display_height,
+		hours,
+		tech,
+		ptr0,
+		len0,
+	);
+	return ret >>> 0;
+}
+
+function getArrayU8FromWasm0(ptr, len) {
+	ptr = ptr >>> 0;
+	return getUint8ArrayMemory0().subarray(ptr / 1, ptr / 1 + len);
+}
+/**
+ * @param {Uint8Array} pixels
+ * @returns {Uint8Array}
+ */
+export function convert_rgba_to_dark_mode(pixels) {
+	const ptr0 = passArray8ToWasm0(pixels, wasm.__wbindgen_malloc);
+	const len0 = WASM_VECTOR_LEN;
+	const ret = wasm.convert_rgba_to_dark_mode(ptr0, len0);
+	var v2 = getArrayU8FromWasm0(ret[0], ret[1]).slice();
+	wasm.__wbindgen_free(ret[0], ret[1] * 1, 1);
+	return v2;
+}
+
+/**
+ * Chroma subsampling format
+ * @enum {0 | 1 | 2 | 3}
+ */
+export const ChromaSampling = Object.freeze({
+	/**
+	 * Both vertically and horizontally subsampled.
+	 */
+	Cs420: 0,
+	0: "Cs420",
+	/**
+	 * Horizontally subsampled.
+	 */
+	Cs422: 1,
+	1: "Cs422",
+	/**
+	 * Not subsampled.
+	 */
+	Cs444: 2,
+	2: "Cs444",
+	/**
+	 * Monochrome.
+	 */
+	Cs400: 3,
+	3: "Cs400",
+});
+/**
+ * @enum {0 | 1}
+ */
+export const DisplayTech = Object.freeze({
+	LCD: 0,
+	0: "LCD",
+	OLED: 1,
+	1: "OLED",
+});
+
+async function __wbg_load(module, imports) {
+	if (typeof Response === "function" && module instanceof Response) {
+		if (typeof WebAssembly.instantiateStreaming === "function") {
+			try {
+				return await WebAssembly.instantiateStreaming(module, imports);
+			} catch (e) {
+				if (module.headers.get("Content-Type") != "application/wasm") {
+					console.warn(
+						"`WebAssembly.instantiateStreaming` failed because your server does not serve Wasm with `application/wasm` MIME type. Falling back to `WebAssembly.instantiate` which is slower. Original error:\n",
+						e,
+					);
+				} else {
+					throw e;
+				}
+			}
+		}
+
+		const bytes = await module.arrayBuffer();
+		return await WebAssembly.instantiate(bytes, imports);
+	} else {
+		const instance = await WebAssembly.instantiate(module, imports);
+
+		if (instance instanceof WebAssembly.Instance) {
+			return { instance, module };
+		} else {
+			return instance;
+		}
+	}
+}
+
+function __wbg_get_imports() {
+	const imports = {};
+	imports.wbg = {};
+	imports.wbg.__wbg_log_c222819a41e063d3 = () =>
+		logError((arg0) => {
+			console.log(arg0);
+		}, arguments);
+	imports.wbg.__wbindgen_init_externref_table = () => {
+		const table = wasm.__wbindgen_export_0;
+		const offset = table.grow(4);
+		table.set(0, undefined);
+		table.set(offset + 0, undefined);
+		table.set(offset + 1, null);
+		table.set(offset + 2, true);
+		table.set(offset + 3, false);
+	};
+	imports.wbg.__wbindgen_string_new = (arg0, arg1) => {
+		const ret = getStringFromWasm0(arg0, arg1);
+		return ret;
+	};
+	imports.wbg.__wbindgen_throw = (arg0, arg1) => {
+		throw new Error(getStringFromWasm0(arg0, arg1));
+	};
+
+	return imports;
+}
+
+function __wbg_init_memory(imports, memory) {}
+
+function __wbg_finalize_init(instance, module) {
+	wasm = instance.exports;
+	__wbg_init.__wbindgen_wasm_module = module;
+	cachedUint8ArrayMemory0 = null;
+
+	wasm.__wbindgen_start();
+	return wasm;
+}
+
+function initSync(module) {
+	if (wasm !== undefined) return wasm;
+
+	if (typeof module !== "undefined") {
+		if (Object.getPrototypeOf(module) === Object.prototype) {
+			({ module } = module);
+		} else {
+			console.warn(
+				"using deprecated parameters for `initSync()`; pass a single object instead",
+			);
+		}
+	}
+
+	const imports = __wbg_get_imports();
+
+	__wbg_init_memory(imports);
+
+	if (!(module instanceof WebAssembly.Module)) {
+		module = new WebAssembly.Module(module);
+	}
+
+	const instance = new WebAssembly.Instance(module, imports);
+
+	return __wbg_finalize_init(instance, module);
+}
+
+async function __wbg_init(module_or_path) {
+	if (wasm !== undefined) return wasm;
+
+	if (typeof module_or_path !== "undefined") {
+		if (Object.getPrototypeOf(module_or_path) === Object.prototype) {
+			({ module_or_path } = module_or_path);
+		} else {
+			console.warn(
+				"using deprecated parameters for the initialization function; pass a single object instead",
+			);
+		}
+	}
+
+	if (typeof module_or_path === "undefined") {
+		module_or_path = new URL("wasm_mod_bg.wasm", import.meta.url);
+	}
+	const imports = __wbg_get_imports();
+
+	if (
+		typeof module_or_path === "string" ||
+		(typeof Request === "function" && module_or_path instanceof Request) ||
+		(typeof URL === "function" && module_or_path instanceof URL)
+	) {
+		module_or_path = fetch(module_or_path);
+	}
+
+	__wbg_init_memory(imports);
+
+	const { instance, module } = await __wbg_load(await module_or_path, imports);
+
+	return __wbg_finalize_init(instance, module);
+}
+
+export { initSync };
+export default __wbg_init;
